fix(value-xs): coerce object-dtype empty fundamentals frame to float64
Zero-row assets (bnb, eos_eth, trx_eth) built their empty parquet with
columns=METRICS and no dtype, inferring object dtype -- crashed np.log
in zscore_signal. Fix at the source: fetch_asset's empty branch now
builds explicit float64 columns; the three affected parquets are
rewritten on disk. Fix defensively: value_ratio coerces its assembled
output to float64 so it doesn't depend on every caller's input frames
being perfectly typed. Regression tests pin both: the empty-frame
round-trip test now asserts dtype, and a new value_ratio test feeds an
empty object-dtype frame alongside a normal one. Both proven to fail
against the pre-fix code.

+++ tradingagents/xsect/value_xs.py
@@ -46,7 +46,7 @@ def value_ratio(fund: dict, metric: str, all_days: pd.DatetimeIndex,
         num = d[METRIC_NUM]
         r = num / den.where(den > 0)
         cols[sym] = r
-    return pd.DataFrame(cols, index=all_days).sort_index(axis=1)
+    return pd.DataFrame(cols, index=all_days).astype("float64").sort_index(axis=1)
 
 
 def zscore_signal(ratio: pd.DataFrame, lag_days: int = 2) -> pd.DataFrame:
